MOdified layout file to fix the exposure of gtag in client side
