+++ package.json
@@ -75,6 +75,7 @@
     "zod": "^3.24.1"
   },
   "devDependencies": {
+    "@playwright/test": "^1.54.1",
     "@types/cors": "^2.8.19",
     "@types/crypto-js": "^4.2.2",
     "@types/next-auth": "^3.13.0",
